+++ cypress/e2e/percorso_completo.cy.ts
@@ -33,6 +33,7 @@ describe('Percorso completo esercizi (M2)', () => {
     cy.get('[data-cy=exercise-stature-max]').type('195')
     cy.get('[data-cy=exercise-youtube]').type('https://www.youtube.com/watch?v=dQw4w9WgXcQ')
     cy.get('[data-cy=video-preview]').should('be.visible')
+    cy.scegliOpzione('exercise-difficulty', 'Media')
     cy.get('[data-cy=exercise-submit]').click()
 
     cy.get('[data-cy=exercise-item]').should('have.length', 2)